requirements.txt
orjson
sanic[ext]
colorama
tomli
pyjwt
py-cpuinfo
sanic-gzip
brotlipy
aiofiles
